fix ogl context problem on mac
