+++ cgogn/rendering/examples/CMakeLists.txt
@@ -12,7 +12,7 @@ set(CGOGN_TEST_MESHES_PATH "${CMAKE_SOURCE_DIR}/data/meshes/")
 add_definitions("-DCGOGN_TEST_MESHES_PATH=${CGOGN_TEST_MESHES_PATH}")
 
 
-add_executable(simple_viewer simple_viewer.cpp main.cpp)
+add_executable(simple_viewer simple_viewer.cpp)
 target_link_libraries(simple_viewer cgogn_core cgogn_rendering QOGLViewer)
