--- cgogn/rendering/examples/main.cpp
@@ -1,56 +1,0 @@
-/****************************************************************************
-
- Copyright (C) 2002-2014 Gilles Debunne. All rights reserved.
-
- This file is part of the QGLViewer library version 2.6.3.
-
- http://www.libqglviewer.com - contact@libqglviewer.com
-
- This file may be used under the terms of the GNU General Public License 
- versions 2.0 or 3.0 as published by the Free Software Foundation and
- appearing in the LICENSE file included in the packaging of this file.
- In addition, as a special exception, Gilles Debunne gives you certain 
- additional rights, described in the file GPL_EXCEPTION in this package.
-
- libQGLViewer uses dual licensing. Commercial/proprietary software must
- purchase a libQGLViewer Commercial License.
-
- This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
- WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
-
-*****************************************************************************/
-
-#include "simple_viewer.h"
-
-
-#include <QApplication>
-
-int main(int argc, char** argv)
-{
-	std::string surfaceMesh;
-	if (argc < 2)
-	{
-		std::cout << "USAGE: " << argv[0] << " [filename]" << std::endl;
-		surfaceMesh = std::string(DEFAULT_MESH_PATH) + std::string("aneurysm3D_1.off");
-		std::cout << "Using default mesh : " << surfaceMesh << std::endl;
-	} else {
-		surfaceMesh = std::string(argv[1]);
-	}
-
-	// Read command lines arguments.
-	QApplication application(argc,argv);
-
-	// Instantiate the viewer.
-	Viewer viewer;
-
-	viewer.setWindowTitle("simpleViewer");
-
-	viewer.import(surfaceMesh);
-
-	// Make the viewer window visible on screen.
-	viewer.show();
-
-//	cgogn::thread_stop();
-	// Run main loop.
-	return application.exec();
-}
